pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "git1file"
version = "0.1.0"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.1",
    "python-multipart>=0.0.6",
    "gitpython>=3.1.40",
    "jinja2>=3.1.2",
]

[tool.setuptools.packages.find]
include = ["git1file*"]